Resolves #5
The main problem with this not working was setting the base directory
for inclusions, especially in SAFE mode, which is what we're using.

This fixes that problem by requiring a new parameter `-include-basedir`
to be set for the doclet to work. I'm also demonstrating this by using
it in the usage section.